+++ lib/main.dart
@@ -12,26 +12,26 @@ import 'notifiers/line_line_statuses_filter_change_notifier.dart';
 import 'notifiers/line_predictions_filter_change_notifier.dart';
 import 'notifiers/lines_filter_change_notifier.dart';
 import 'notifiers/tfl_api_change_notifier.dart';
+import 'pages/lines/line_disruption_page.dart';
+import 'pages/lines/line_line_disruptions_page.dart';
+import 'pages/lines/line_line_routes_page.dart';
+import 'pages/lines/line_line_statuses_page.dart';
+import 'pages/lines/line_page.dart';
+import 'pages/lines/line_predictions_page.dart';
+import 'pages/lines/line_route_page.dart';
+import 'pages/lines/line_route_sequences_page.dart';
+import 'pages/lines/line_status_page.dart';
+import 'pages/lines/line_stop_points_page.dart';
+import 'pages/lines/lines_page.dart';
+import 'pages/predictions/prediction_page.dart';
+import 'pages/route_sequences/route_sequence_page.dart';
+import 'pages/route_sequences/route_sequence_stop_point_sequences_page.dart';
+import 'pages/settings/settings_page.dart';
+import 'pages/stop_points/stop_point_page.dart';
+import 'pages/stop_point_sequences/stop_point_sequence_page.dart';
+import 'pages/stop_point_sequences/stop_point_sequence_stop_points_page.dart';
 import 'pages/home_page.dart';
-import 'pages/line_disruption_page.dart';
-import 'pages/line_line_disruptions_page.dart';
-import 'pages/line_line_routes_page.dart';
-import 'pages/line_line_statuses_page.dart';
-import 'pages/line_page.dart';
-import 'pages/line_predictions_page.dart';
-import 'pages/line_route_page.dart';
-import 'pages/line_route_sequences_page.dart';
-import 'pages/line_status_page.dart';
-import 'pages/line_stop_points_page.dart';
-import 'pages/lines_page.dart';
 import 'pages/login_page.dart';
-import 'pages/prediction_page.dart';
-import 'pages/route_sequence_page.dart';
-import 'pages/route_sequence_stop_point_sequences_page.dart';
-import 'pages/settings_page.dart';
-import 'pages/stop_point_page.dart';
-import 'pages/stop_point_sequence_page.dart';
-import 'pages/stop_point_sequence_stop_points_page.dart';
 
 Future<void> main() async {
   await _configureIntl();
