+++ lib/pages/lines/line_line_disruptions_page.dart
@@ -4,58 +4,59 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 
-import '../material/list_tile.dart';
-import '../notifiers/tfl_api_change_notifier.dart';
-import '../widgets/async.dart';
+import '../../material/list_tile.dart';
+import '../../notifiers/tfl_api_change_notifier.dart';
+import '../../widgets/async.dart';
 
-class LineRouteSequencesPage extends StatefulWidget {
-  static const route = '/lines/:id/route_sequences';
+class LineLineDisruptionsPage extends StatefulWidget {
+  static const route = '/lines/:id/line_disruptions';
 
   final Line line;
 
-  LineRouteSequencesPage({
+  LineLineDisruptionsPage({
     Key key,
     @required this.line,
   }) : super(key: key);
 
   @override
-  _LineRouteSequencesPageState createState() => _LineRouteSequencesPageState();
+  _LineLineDisruptionsPageState createState() =>
+      _LineLineDisruptionsPageState();
 }
 
-class _LineRouteSequencesPageState extends State<LineRouteSequencesPage> {
-  StreamController<List<RouteSequence>> _streamController;
+class _LineLineDisruptionsPageState extends State<LineLineDisruptionsPage> {
+  StreamController<List<LineDisruption>> _streamController;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Route sequences'),
+        title: Text('Line disruptions'),
       ),
       body: Consumer<TflApiChangeNotifier>(
         builder: (context, tflApi, child) {
-          final getRouteSequences = () {
-            return tflApi.tflApi.lines.getRouteSequences(widget.line.id);
+          final getLineDisruptions = () {
+            return tflApi.tflApi.lines.getLineDisruptions(widget.line.id);
           };
 
-          getRouteSequences()
+          getLineDisruptions()
               .then(_streamController.add)
               .catchError(_streamController.addError);
 
-          return CircularProgressIndicatorStreamBuilder<List<RouteSequence>>(
+          return CircularProgressIndicatorStreamBuilder<List<LineDisruption>>(
             stream: _streamController.stream,
             builder: (context, data) {
               return RefreshIndicator(
                 child: ListView.builder(
                   itemBuilder: (context, index) {
-                    return RouteSequenceListTile(
+                    return LineDisruptionListTile(
                       context: context,
-                      routeSequence: data[index],
+                      lineDisruption: data[index],
                     );
                   },
                   itemCount: data.length,
                 ),
                 onRefresh: () {
-                  return getRouteSequences()
+                  return getLineDisruptions()
                       .then(_streamController.add)
                       .catchError(_streamController.addError);
                 },
@@ -78,6 +79,6 @@ class _LineRouteSequencesPageState extends State<LineRouteSequencesPage> {
   void initState() {
     super.initState();
 
-    _streamController = StreamController<List<RouteSequence>>();
+    _streamController = StreamController<List<LineDisruption>>();
   }
 }
