+++ lib/pages/lines/line_line_statuses_page.dart
@@ -4,34 +4,34 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 
-import '../material/list_tile.dart';
-import '../notifiers/line_predictions_filter_change_notifier.dart';
-import '../notifiers/tfl_api_change_notifier.dart';
-import '../widgets/async.dart';
-import 'line_predictions_filter_page.dart';
+import '../../material/list_tile.dart';
+import '../../notifiers/line_line_statuses_filter_change_notifier.dart';
+import '../../notifiers/tfl_api_change_notifier.dart';
+import '../../widgets/async.dart';
+import 'line_line_statuses_filter_page.dart';
 
-class LinePredictionsPage extends StatefulWidget {
-  static const route = '/lines/:id/predictions';
+class LineLineStatusesPage extends StatefulWidget {
+  static const route = '/lines/:id/line_statuses';
 
   final Line line;
 
-  LinePredictionsPage({
+  LineLineStatusesPage({
     Key key,
     @required this.line,
   }) : super(key: key);
 
   @override
-  _LinePredictionsPageState createState() => _LinePredictionsPageState();
+  _LineLineStatusesPageState createState() => _LineLineStatusesPageState();
 }
 
-class _LinePredictionsPageState extends State<LinePredictionsPage> {
-  StreamController<List<Prediction>> _streamController;
+class _LineLineStatusesPageState extends State<LineLineStatusesPage> {
+  StreamController<List<LineStatus>> _streamController;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Predictions'),
+        title: Text('Line statuses'),
         actions: <Widget>[
           IconButton(
             icon: Icon(Icons.filter_list),
@@ -39,7 +39,7 @@ class _LinePredictionsPageState extends State<LinePredictionsPage> {
               Navigator.of(context).push(
                 MaterialPageRoute(
                   builder: (context) {
-                    return LinePredictionsFilterPage(line: widget.line);
+                    return LineLineStatusesFilterPage();
                   },
                   fullscreenDialog: true,
                 ),
@@ -50,35 +50,37 @@ class _LinePredictionsPageState extends State<LinePredictionsPage> {
       ),
       body: Consumer<TflApiChangeNotifier>(
         builder: (context, tflApi, child) {
-          return Consumer<LinePredictionsFilterChangeNotifier>(
-            builder: (context, linePredictionsFilter, child) {
-              final getPredictions = () {
-                return tflApi.tflApi.lines.getPredictions(
+          return Consumer<LineLineStatusesFilterChangeNotifier>(
+            builder: (context, lineLineStatusesFilter, child) {
+              final getLineStatuses = () {
+                return tflApi.tflApi.lines.getLineStatuses(
                   widget.line.id,
-                  stopPointId: linePredictionsFilter.stopPoint?.id,
-                  destinationStationId: linePredictionsFilter.destination?.id,
+                  startDate: lineLineStatusesFilter.date?.toIso8601String(),
+                  endDate: lineLineStatusesFilter.date
+                      ?.add(Duration(days: 1))
+                      ?.toIso8601String(),
                 );
               };
 
-              getPredictions()
+              getLineStatuses()
                   .then(_streamController.add)
                   .catchError(_streamController.addError);
 
-              return CircularProgressIndicatorStreamBuilder<List<Prediction>>(
+              return CircularProgressIndicatorStreamBuilder<List<LineStatus>>(
                 stream: _streamController.stream,
                 builder: (context, data) {
                   return RefreshIndicator(
                     child: ListView.builder(
                       itemBuilder: (context, index) {
-                        return PredictionListTile(
+                        return LineStatusListTile(
                           context: context,
-                          prediction: data[index],
+                          lineStatus: data[index],
                         );
                       },
                       itemCount: data.length,
                     ),
                     onRefresh: () {
-                      return getPredictions()
+                      return getLineStatuses()
                           .then(_streamController.add)
                           .catchError(_streamController.addError);
                     },
@@ -103,6 +105,6 @@ class _LinePredictionsPageState extends State<LinePredictionsPage> {
   void initState() {
     super.initState();
 
-    _streamController = StreamController<List<Prediction>>();
+    _streamController = StreamController<List<LineStatus>>();
   }
 }
