+++ lib/pages/lines/line_route_sequences_page.dart
@@ -4,58 +4,58 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 
-import '../material/list_tile.dart';
-import '../notifiers/tfl_api_change_notifier.dart';
-import '../widgets/async.dart';
+import '../../material/list_tile.dart';
+import '../../notifiers/tfl_api_change_notifier.dart';
+import '../../widgets/async.dart';
 
-class LineStopPointsPage extends StatefulWidget {
-  static const route = '/lines/:id/stop_points';
+class LineRouteSequencesPage extends StatefulWidget {
+  static const route = '/lines/:id/route_sequences';
 
   final Line line;
 
-  LineStopPointsPage({
+  LineRouteSequencesPage({
     Key key,
     @required this.line,
   }) : super(key: key);
 
   @override
-  _LineStopPointsPageState createState() => _LineStopPointsPageState();
+  _LineRouteSequencesPageState createState() => _LineRouteSequencesPageState();
 }
 
-class _LineStopPointsPageState extends State<LineStopPointsPage> {
-  StreamController<List<StopPoint>> _streamController;
+class _LineRouteSequencesPageState extends State<LineRouteSequencesPage> {
+  StreamController<List<RouteSequence>> _streamController;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Stop points'),
+        title: Text('Route sequences'),
       ),
       body: Consumer<TflApiChangeNotifier>(
         builder: (context, tflApi, child) {
-          final getStopPoints = () {
-            return tflApi.tflApi.lines.getStopPoints(widget.line.id);
+          final getRouteSequences = () {
+            return tflApi.tflApi.lines.getRouteSequences(widget.line.id);
           };
 
-          getStopPoints()
+          getRouteSequences()
               .then(_streamController.add)
               .catchError(_streamController.addError);
 
-          return CircularProgressIndicatorStreamBuilder<List<StopPoint>>(
+          return CircularProgressIndicatorStreamBuilder<List<RouteSequence>>(
             stream: _streamController.stream,
             builder: (context, data) {
               return RefreshIndicator(
                 child: ListView.builder(
                   itemBuilder: (context, index) {
-                    return StopPointListTile(
+                    return RouteSequenceListTile(
                       context: context,
-                      stopPoint: data[index],
+                      routeSequence: data[index],
                     );
                   },
                   itemCount: data.length,
                 ),
                 onRefresh: () {
-                  return getStopPoints()
+                  return getRouteSequences()
                       .then(_streamController.add)
                       .catchError(_streamController.addError);
                 },
@@ -78,6 +78,6 @@ class _LineStopPointsPageState extends State<LineStopPointsPage> {
   void initState() {
     super.initState();
 
-    _streamController = StreamController<List<StopPoint>>();
+    _streamController = StreamController<List<RouteSequence>>();
   }
 }
