+++ lib/pages/lines/lines_filter_page.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 
-import '../material/radio_list_tile.dart';
-import '../notifiers/lines_filter_change_notifier.dart';
-import '../notifiers/tfl_api_change_notifier.dart';
-import '../widgets/async.dart';
+import '../../material/radio_list_tile.dart';
+import '../../notifiers/lines_filter_change_notifier.dart';
+import '../../notifiers/tfl_api_change_notifier.dart';
+import '../../widgets/async.dart';
 
 class LinesFilterPage extends StatefulWidget {
   LinesFilterPage({Key key}) : super(key: key);
