+++ lib/pages/lines/lines_page.dart
@@ -4,34 +4,30 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 
-import '../material/list_tile.dart';
-import '../notifiers/line_line_statuses_filter_change_notifier.dart';
-import '../notifiers/tfl_api_change_notifier.dart';
-import '../widgets/async.dart';
-import 'line_line_statuses_filter_page.dart';
+import '../../material/list_tile.dart';
+import '../../notifiers/lines_filter_change_notifier.dart';
+import '../../notifiers/tfl_api_change_notifier.dart';
+import '../../widgets/async.dart';
+import '../../widgets/drawer.dart';
+import 'lines_filter_page.dart';
 
-class LineLineStatusesPage extends StatefulWidget {
-  static const route = '/lines/:id/line_statuses';
+class LinesPage extends StatefulWidget {
+  static const route = '/lines';
 
-  final Line line;
-
-  LineLineStatusesPage({
-    Key key,
-    @required this.line,
-  }) : super(key: key);
+  LinesPage({Key key}) : super(key: key);
 
   @override
-  _LineLineStatusesPageState createState() => _LineLineStatusesPageState();
+  _LinesPageState createState() => _LinesPageState();
 }
 
-class _LineLineStatusesPageState extends State<LineLineStatusesPage> {
-  StreamController<List<LineStatus>> _streamController;
+class _LinesPageState extends State<LinesPage> {
+  StreamController<List<Line>> _streamController;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Line statuses'),
+        title: Text('Lines'),
         actions: <Widget>[
           IconButton(
             icon: Icon(Icons.filter_list),
@@ -39,7 +35,7 @@ class _LineLineStatusesPageState extends State<LineLineStatusesPage> {
               Navigator.of(context).push(
                 MaterialPageRoute(
                   builder: (context) {
-                    return LineLineStatusesFilterPage();
+                    return LinesFilterPage();
                   },
                   fullscreenDialog: true,
                 ),
@@ -50,37 +46,33 @@ class _LineLineStatusesPageState extends State<LineLineStatusesPage> {
       ),
       body: Consumer<TflApiChangeNotifier>(
         builder: (context, tflApi, child) {
-          return Consumer<LineLineStatusesFilterChangeNotifier>(
-            builder: (context, lineLineStatusesFilter, child) {
-              final getLineStatuses = () {
-                return tflApi.tflApi.lines.getLineStatuses(
-                  widget.line.id,
-                  startDate: lineLineStatusesFilter.date?.toIso8601String(),
-                  endDate: lineLineStatusesFilter.date
-                      ?.add(Duration(days: 1))
-                      ?.toIso8601String(),
+          return Consumer<LinesFilterChangeNotifier>(
+            builder: (context, linesFilter, child) {
+              final getLines = () {
+                return tflApi.tflApi.lines.get(
+                  mode: linesFilter.mode?.modeName,
                 );
               };
 
-              getLineStatuses()
+              getLines()
                   .then(_streamController.add)
                   .catchError(_streamController.addError);
 
-              return CircularProgressIndicatorStreamBuilder<List<LineStatus>>(
+              return CircularProgressIndicatorStreamBuilder<List<Line>>(
                 stream: _streamController.stream,
                 builder: (context, data) {
                   return RefreshIndicator(
                     child: ListView.builder(
                       itemBuilder: (context, index) {
-                        return LineStatusListTile(
+                        return LineListTile(
                           context: context,
-                          lineStatus: data[index],
+                          line: data[index],
                         );
                       },
                       itemCount: data.length,
                     ),
                     onRefresh: () {
-                      return getLineStatuses()
+                      return getLines()
                           .then(_streamController.add)
                           .catchError(_streamController.addError);
                     },
@@ -91,6 +83,7 @@ class _LineLineStatusesPageState extends State<LineLineStatusesPage> {
           );
         },
       ),
+      drawer: AppDrawer(),
     );
   }
 
@@ -105,6 +98,6 @@ class _LineLineStatusesPageState extends State<LineLineStatusesPage> {
   void initState() {
     super.initState();
 
-    _streamController = StreamController<List<LineStatus>>();
+    _streamController = StreamController<List<Line>>();
   }
 }
