+++ lib/pages/route_sequences/route_sequence_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 
-import '../material/scroll_view.dart';
+import '../../material/scroll_view.dart';
 
 class RouteSequencePage extends StatefulWidget {
   static const route = '/route_sequences/:id';
