+++ lib/pages/route_sequences/route_sequence_stop_point_sequences_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 
-import '../material/list_tile.dart';
+import '../../material/list_tile.dart';
 
 class RouteSequenceStopPointSequencesPage extends StatefulWidget {
   static const route = '/route_sequences/:id/stop_point_sequences';
