+++ lib/pages/stop_points/stop_point_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 
-import '../material/scroll_view.dart';
+import '../../material/scroll_view.dart';
 
 class StopPointPage extends StatefulWidget {
   static const route = '/stop_points/:id';
